HW02: Конфигурирование приложение в Java-based стиле. Рефакторинг: переименование пакета.
