+++ hw01-intro/src/main/java/ru/otus/hw02/dao/exception/QuestionLoadingException.java
