+++ hw01-intro/src/main/java/ru/otus/hw02/domain/Answer.java
