+++ hw01-intro/src/main/java/ru/otus/hw02/service/QuestionConsolePrintServiceImpl.java
