+++ hw01-intro/src/test/java/ru/otus/hw02/domain/QuestionTest.java
