+++ hw02-annotation-based/src/main/java/ru/otus/hw02/dao/AnswerDao.java
@@ -1,6 +1,6 @@
-package ru.otus.hw01.dao;
+package ru.otus.hw02.dao;
 
-import ru.otus.hw01.domain.Answer;
+import ru.otus.hw02.domain.Answer;
 
 import java.util.List;
 import java.util.Optional;
