+++ hw02-annotation-based/src/main/java/ru/otus/hw02/dao/exception/AnswerLoadingException.java
@@ -1,4 +1,4 @@
-package ru.otus.hw01.dao.exception;
+package ru.otus.hw02.dao.exception;
 
 import java.io.IOException;
