+++ hw02-annotation-based/src/main/java/ru/otus/hw02/service/QuestionService.java
@@ -1,6 +1,6 @@
-package ru.otus.hw01.service;
+package ru.otus.hw02.service;
 
-import ru.otus.hw01.domain.Question;
+import ru.otus.hw02.domain.Question;
 
 import java.util.List;
 import java.util.Optional;
